+++ backend/.gitignore
@@ -40,5 +40,5 @@ out/
 db_dev.mv.db
 db_dev.trace.db
 **/application-secret.yml
-api-docs.json
+**/api-docs.json
 package-lock.json
