+++ backend/src/main/java/com/ll/TeamProject/domain/user/controller/AdminAccountController.java
@@ -1,6 +1,7 @@
 package com.ll.TeamProject.domain.user.controller;
 
 import com.ll.TeamProject.domain.user.dto.admin.PasswordChangeRequest;
+import com.ll.TeamProject.domain.user.service.AccountVerificationService;
 import com.ll.TeamProject.domain.user.service.UserService;
 import com.ll.TeamProject.global.rsData.RsData;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -18,12 +19,13 @@ import org.springframework.web.bind.annotation.*;
 public class AdminAccountController {
 
     private final UserService userService;
+    private final AccountVerificationService accountVerificationService;
 
     @PutMapping("/{username}/password")
     public ResponseEntity<RsData<Void>> changePassword(
             @PathVariable("username") String username,
             @RequestBody @Valid PasswordChangeRequest request) {
-        userService.changePassword(username, request.password());
+        accountVerificationService.changePassword(username, request.password());
         return ResponseEntity.ok(new RsData<>("200-1", "비밀번호 변경이 완료되었습니다."));
     }
