+++ backend/src/main/java/com/ll/TeamProject/domain/user/service/UserService.java
@@ -1,157 +1,30 @@
 package com.ll.TeamProject.domain.user.service;
 
 import com.ll.TeamProject.domain.user.dto.UserDto;
-import com.ll.TeamProject.domain.user.dto.admin.LoginDto;
-import com.ll.TeamProject.domain.user.entity.Authentication;
 import com.ll.TeamProject.domain.user.entity.SiteUser;
-import com.ll.TeamProject.domain.user.enums.AuthType;
 import com.ll.TeamProject.domain.user.enums.Role;
-import com.ll.TeamProject.domain.user.repository.AuthenticationRepository;
 import com.ll.TeamProject.domain.user.repository.UserRepository;
 import com.ll.TeamProject.global.exceptions.ServiceException;
-import com.ll.TeamProject.global.mail.EmailService;
 import com.ll.TeamProject.global.userContext.UserContext;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-
-import static com.ll.TeamProject.domain.user.enums.Role.USER;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
-    private static final String VERIFICATION_CODE_KEY = "verificationCode:";
-    private static final String PASSWORD_RESET_KEY = "password-reset:";
 
     private final UserRepository userRepository;
     private final AuthTokenService authTokenService;
-    private final AuthenticationRepository authenticationRepository;
     private final UserContext userContext;
-    private final AuthenticationService authenticationService;
     private final ForbiddenService forbiddenService;
-    private final EmailService emailService;
-    private final StringRedisTemplate redisTemplate;
-    private final PasswordEncoder passwordEncoder;
-
-    public LoginDto login(String username, String password) {
-        SiteUser user = findByUsername(username)
-                .orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 사용자입니다."));
-
-        if (user.isLocked()) throw new ServiceException("403-2", "계정이 잠겨있습니다.");
-
-        if (!passwordEncoder.matches(password, user.getPassword())) {
-            handleFailedLogin(user);
-        }
-
-        return createLoginResponse(user);
-    }
-
-    private void handleFailedLogin(SiteUser user) {
-        authenticationService.handleLoginFailure(user);
-        throw new ServiceException("401-2", "비밀번호가 일치하지 않습니다.");
-    }
-
-    private LoginDto createLoginResponse(SiteUser user) {
-        authenticationService.modifyLastLogin(user);
-        String accessToken = userContext.makeAuthCookies(user);
-        return new LoginDto(new UserDto(user), user.getApiKey(), accessToken);
-    }
-  
-    public void logout(HttpServletRequest request) {
-        request.getSession().invalidate();
-
-        userContext.deleteCookie("accessToken");
-        userContext.deleteCookie("apiKey");
-        userContext.deleteCookie("JSESSIONID");
-
-        SecurityContextHolder.clearContext();
-    }
-
-    public void processVerification(String username, String email) {
-        SiteUser user = validateUsernameAndEmail(username, email);
-
-        String code = generateVerificationCode();
-
-        String redisKey = VERIFICATION_CODE_KEY + username;
-
-        redisTemplate.opsForValue().set(redisKey, code, 180, TimeUnit.SECONDS);
-
-        sendVerificationEmail(user.getNickname(), user.getEmail(), code);
-    }
-
-    public void verifyAndUnlockAccount(String username, String verificationCode) {
-        String redisKey = VERIFICATION_CODE_KEY + username;
-        String storedCode = redisTemplate.opsForValue().get(redisKey);
-
-        if (!verificationCode.equals(storedCode)) {
-            throw new ServiceException("401-3", "인증번호가 유효하지 않거나 만료되었습니다.");
-        }
-
-        redisTemplate.delete(redisKey);
-        redisTemplate.opsForValue().set(PASSWORD_RESET_KEY + username, username, 300, TimeUnit.SECONDS);
-    }
-
-    private SiteUser validateUsernameAndEmail(String username, String email) {
-        return userRepository.findByUsername(username)
-                .map(user -> {
-                    if (!user.getEmail().equals(email)) {
-                        throw new ServiceException("401-2", "이메일이 일치하지 않습니다.");
-                    }
-                    return user;
-                })
-                .orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 사용자입니다."));
-    }
-
-    private String generateVerificationCode() {
-        return UUID.randomUUID().toString().substring(0, 6);
-    }
-
-    private void sendVerificationEmail(String nickname, String email, String verificationCode) {
-        String subject = "계정 인증번호";
-        String content = String.format("안녕하세요, %s님.\n\n인증번호: %s\n인증번호는 3분 동안 유효합니다.", nickname, verificationCode);
-
-        emailService.sendEmail(email, subject, content);
-    }
-
-    private void unlockAccount(SiteUser user) {
-        user.unlockAccount();
-        userRepository.save(user);
-    }
-
-    public void unlockAccount(Long id) {
-        SiteUser user = findById(id).get();
-        user.unlockAccount();
-        userRepository.save(user);
-    }
-
-    public void changePassword(String username, String password) {
-        String redisKey = PASSWORD_RESET_KEY + username;
-        String storedUsername = redisTemplate.opsForValue().get(redisKey);
-
-        if (!username.equals(storedUsername)) {
-            redisTemplate.delete(redisKey);
-            throw new ServiceException("401-3", "올바른 요청이 아닙니다.");
-        }
-
-        SiteUser user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 사용자입니다."));
-
-        user.changePassword(passwordEncoder.encode(password));
-        unlockAccount(user);
-        userRepository.save(user);
-        redisTemplate.delete(redisKey);
-    }
+    private final AuthService authService;
 
     public Optional<SiteUser> findByUsername(String username) {
         return userRepository.findByUsername(username);
@@ -188,57 +61,6 @@ public class UserService {
         return userRepository.findByRoleAndIsDeletedFalse(role, pageRequest);
     }
 
-    public String genAccessToken(SiteUser user) {
-        return authTokenService.genAccessToken(user);
-    }
-
-    public String genAuthToken(SiteUser user) {
-        return user.getApiKey() + " " + genAccessToken(user);
-    }
-
-    // JWT 로 얻은 가짜 user 객체 (DB 에서 조회한 user 아님)
-    public SiteUser getUserFromAccessToken(String accessToken) {
-        Map<String, Object> payload = authTokenService.payload(accessToken);
-
-        if (payload == null) return null;
-
-        long id = (long) payload.get("id");
-        String username = (String) payload.get("username");
-        String nickname = (String) payload.get("nickname");
-        Role role = (Role) payload.get("role");
-
-        return new SiteUser(id, username, nickname, role);
-    }
-
-    public SiteUser findOrRegisterUser(String username, String email, String providerTypeCode) {
-        return findByUsername(username)
-                .orElseGet(() -> join(username, "", email, providerTypeCode));
-    }
-
-    public SiteUser join(String username, String password, String email, String providerTypeCode) {
-        SiteUser user = SiteUser.builder()
-                .username(username)
-                .password(password)
-                .nickname(username)
-                .email(email)
-                .role(USER)
-                .apiKey(UUID.randomUUID().toString())
-                .locked(false)
-                .build();
-        user = userRepository.save(user);
-
-        AuthType authType = AuthType.valueOf(providerTypeCode);
-        Authentication authentication = Authentication
-                .builder()
-                .user(user)
-                .authType(authType)
-                .failedAttempts(0)
-                .build();
-        authenticationRepository.save(authentication);
-
-        return user;
-    }
-
     public void modify(String nickname) {
         if (forbiddenService.isForbidden(nickname)) {
             throw new ServiceException("400-1", "해당 닉네임은 사용할 수 없습니다.");
@@ -268,7 +90,12 @@ public class UserService {
         userToDelete.delete();
         userRepository.save(userToDelete);
 
-        logout(request);
+        // TODO: 로그아웃 추가 수정 필요
+        userContext.deleteCookie("accessToken");
+        userContext.deleteCookie("apiKey");
+        userContext.deleteCookie("JSESSIONID");
+
+        authService.logout();
         return new UserDto(userToDelete);
     }
 
@@ -279,5 +106,33 @@ public class UserService {
         if (!userToDelete.getUsername().equals(actor.getUsername())) {
             throw new ServiceException("403-1", "접근 권한이 없습니다.");
         }
+    }
+
+    public void unlockAccount(Long id) {
+        SiteUser user = findById(id).get();
+        user.unlockAccount();
+        userRepository.save(user);
+    }
+
+    public String genAccessToken(SiteUser user) {
+        return authTokenService.genAccessToken(user);
+    }
+
+    public String genAuthToken(SiteUser user) {
+        return user.getApiKey() + " " + genAccessToken(user);
+    }
+
+    // JWT 로 얻은 가짜 user 객체 (DB 에서 조회한 user 아님)
+    public SiteUser getUserFromAccessToken(String accessToken) {
+        Map<String, Object> payload = authTokenService.payload(accessToken);
+
+        if (payload == null) return null;
+
+        long id = (long) payload.get("id");
+        String username = (String) payload.get("username");
+        String nickname = (String) payload.get("nickname");
+        Role role = (Role) payload.get("role");
+
+        return new SiteUser(id, username, nickname, role);
     }
 }
