+++ backend/src/main/java/com/ll/TeamProject/global/redis/RedisService.java
@@ -1,0 +1,26 @@
+package com.ll.TeamProject.global.redis;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+
+@Service
+@RequiredArgsConstructor
+public class RedisService {
+    private final StringRedisTemplate redisTemplate;
+
+    public void setValue(String key, String value, long timeoutSeconds) {
+        redisTemplate.opsForValue().set(key, value, timeoutSeconds, TimeUnit.SECONDS);
+    }
+
+    public Optional<String> getValue(String key) {
+        return Optional.ofNullable(redisTemplate.opsForValue().get(key));
+    }
+
+    public void deleteValue(String key) {
+        redisTemplate.delete(key);
+    }
+}
