+++ backend/src/main/java/com/ll/TeamProject/global/security/CustomOAuth2UserService.java
@@ -2,7 +2,7 @@ package com.ll.TeamProject.global.security;
 
 import com.ll.TeamProject.domain.user.entity.SiteUser;
 import com.ll.TeamProject.domain.user.service.AuthenticationService;
-import com.ll.TeamProject.domain.user.service.UserService;
+import com.ll.TeamProject.domain.user.service.JoinService;
 import com.ll.TeamProject.global.userContext.UserContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
@@ -18,9 +18,9 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
-    private final UserService userService;
     private final AuthenticationService authenticationService;
     private final UserContext userContext;
+    private final JoinService joinService;
 
     // 소셜 로그인이 성공할 때마다 이 함수가 실행된다.
     @Transactional
@@ -55,7 +55,7 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
             email = accountProperties.get("email");
         }
 
-        SiteUser user = userService.findOrRegisterUser(username, email, providerTypeCode);
+        SiteUser user = joinService.findOrRegisterUser(username, email, providerTypeCode);
 
         authenticationService.modifyLastLogin(user);
         userContext.setLongCookie("lastLogin", providerTypeCode);
